Fill NA gaps in temperature data for resampling
I do not get temperature data after resampling. I believe the reason is that I have long NA gaps in the temperature column because of the different sampling frequency in the sensors. I try to fill the NA gaps with the last observartion to see if the resampled data include temperature values.